format: consolidate json and asn1 under a single subdir
Both subsystems are structured-data encoding formats; grouping them
under format/ leaves room for future xml / yaml / msgpack / cbor
codecs without sprinkling more single-format subdirs at the top
level.

The asn1/ subdir + root rasn1.h umbrella collapse into a flat
format/ — rlib.h pulls in format/rasn1.h and format/roid.h
directly, matching the existing convention for the json files.
The @defgroup r_asn1 block migrates from the removed umbrella
into format/rasn1.h so the Doxygen Topic structure is unchanged.
