--- include/rlib/rasn1.h
@@ -1,40 +1,0 @@
-/* RLIB - Convenience library for useful things
- * Copyright (C) 2016-2017 Haakon Sporsheim <haakon.sporsheim@gmail.com>
- *
- * This library is free software; you can redistribute it and/or
- * modify it under the terms of the GNU Lesser General Public
- * License as published by the Free Software Foundation; either
- * version 3.0 of the License, or (at your option) any later version.
- *
- * This library is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
- * Lesser General Public License for more details.
- *
- * You should have received a copy of the GNU Lesser General Public
- * License along with this library.
- * See the COPYING file at the root of the source repository.
- */
-#ifndef __R_ASN1_H__
-#define __R_ASN1_H__
-
-/**
- * @defgroup r_asn1 ASN.1 BER / DER
- *
- * @brief Encoder and decoder for ASN.1 in Basic and Distinguished
- * Encoding Rules, plus the object-identifier (OID) registry that
- * goes with them.
- *
- * Consumed heavily by @c r_crypto (X.509 certificates, PKCS#1 / #8
- * key formats, PEM bodies) but the encoding itself is general -
- * SNMP, LDAP, OCSP and many other on-disk and on-wire formats use
- * the same primitives.
- */
-
-#include <rlib/rlib.h>
-
-#include <rlib/asn1/rasn1.h>
-#include <rlib/asn1/roid.h>
-
-#endif /* __R_ASN1_H__ */
-
