+++ rlib/crypto/rx509.c
@@ -20,7 +20,7 @@
 #include "rcrypto-private.h"
 #include <rlib/crypto/rx509.h>
 
-#include <rlib/asn1/roid.h>
+#include <rlib/format/roid.h>
 #include <rlib/data/rlist.h>
 
 #include <rlib/rmem.h>
