+++ rlib/format/rasn1-private.h
@@ -24,7 +24,7 @@
 
 #include <rlib/rtypes.h>
 
-#include <rlib/asn1/rasn1.h>
+#include <rlib/format/rasn1.h>
 #include <rlib/data/rlist.h>
 #include <rlib/rmemfile.h>
