+++ rlib/format/rber.c
