+++ rlib/meson.build
@@ -11,8 +11,8 @@ rlib_sources = [
   'os/renv.c',
   'rio.c',
   'net/riosocket.c',
-  'rjson.c',
-  'rjsonparser.c',
+  'format/rjson.c',
+  'format/rjsonparser.c',
   'rlibinit.c',
   'rlog.c',
   'rmath.c',
@@ -38,12 +38,12 @@ rlib_sources = [
   'rtime.c',
   'os/rtty.c',
   'ruri.c',
-  'asn1/rasn1bin.c',
-  'asn1/rasn1binenc.c',
-  'asn1/rasn1.c',
-  'asn1/rber.c',
-  'asn1/rder.c',
-  'asn1/roid.c',
+  'format/rasn1bin.c',
+  'format/rasn1binenc.c',
+  'format/rasn1.c',
+  'format/rber.c',
+  'format/rder.c',
+  'format/roid.c',
   'binfmt/relfparser.c',
   'binfmt/rmacho.c',
   'binfmt/rmachoparser.c',
